Adding docstrings

+++ complete_python_bootcamp/projects/blackjack/player.py
@@ -1,8 +1,8 @@
 class Player:
-    def __init__(self, total_amount: int, points: int = 0):
+    def __init__(self, total_amount: float, points: int = 0):
         # Player points
         self.points: int = points
         # Players total amount of money
-        self.total_amount: int = total_amount
+        self.total_amount: float = total_amount
         # Player hand
         self.hand: list = []
